requirements.txt
pandas
telethon==1.37.0
PyYAML
openpyxl